+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Write the date in place of the "Unreleased" in the case a new version is release
 
 - `docker-compose.yml` now uses the healthcheck endpoint `/healthz`
 - Fix bug where access policies were not applied to child nodes during request
+- Add metadata-based access control to SimpleAccessPolicy
+- Add example test of metadata-based allowed_scopes which requires the path to the target node
 
 ## 2024-12-09